// generated/src/app/api-models/vendor-adv-pay-rules.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class VendorAdvPayRules {
  advPayType: string;
  replenish: string;
  offsetBy: string;
  zeroPayOption: string;
  accPayType: string;
  userDefined1: string;
  userDefined2: string;
  userDefined3: string;
  userDefined4: string;
  userDate1: Date;
  userDate2: Date;
  userDate3: Date;
  userDate4: Date;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}